Refactor and fix circular import for config
